Prepare Order: Sync Stock value from DB; Add txt_PassMessage;
Stock function is not all ready.

+++ VNW/Views/ViewModels/ShoppingCart.cs
@@ -12,6 +12,6 @@ namespace VNW.ViewModels
         public int Price;
         public string Img;
         public string Name;
-        public int Stock;
+        public short Stock;
     }
 }
